+++ src/stories/Button.stories.tsx
@@ -1,29 +1,33 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 
-import Button from '../components/Button'
+import CardButton from '/@/components/CardButton'
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'componets/Button',
-  component: Button
-} as ComponentMeta<typeof Button>
+  title: 'componets/CardButton',
+  component: CardButton,
+} as ComponentMeta<typeof CardButton>
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />
+const Template: ComponentStory<typeof CardButton> = (args) => (
+  <CardButton {...args} />
+)
 
 export const HasBackground = Template.bind({})
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 HasBackground.args = {
-  icon: <img src="src/assets/react.svg" />,
+  icon: <img src="src/assets/react.svg" alt="" />,
   title: 'title',
   description: 'description',
-  onClick: () => { console.log('clicked!') },
-  hasBackground: true
+  onClick: () => {
+    console.log('clicked!')
+  },
+  hasBackground: true,
 }
 
 export const NoBackground = Template.bind({})
 NoBackground.args = {
   ...HasBackground.args,
-  hasBackground: false
+  hasBackground: false,
 }
